Renamed Highlights to ArenaTv
Added a sort of pipeline to change how the array and object looks when passed to playlist.
Added a onPlaylistAtEnd function that is passed along as a prop to the Videoplayer.
Added jsx for VideoPlayer component

Modified so that the pipeline checks current node environment, big buck bunny is the bomb.